+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of the syslog_ng cookbook.
 
+## 0.3.1 (2019-07-09)
+
+- Add package exclude option to install resource
+- Add `githead` installation option
+
 ## 0.3.0 (2019-07-04)
 
 - Default COPR repository updated to v3.22